refactor: create secret manager
