+++ internal/swarm/swarm.go
@@ -3,11 +3,13 @@ package swarm
 import "github.com/docker/docker/client"
 
 type Swarm struct {
-	ServiceManager *ServiceManager
+	Services *ServiceManager
+	Secrets  *SecretManager
 }
 
 func NewSwarm(dockerClient *client.Client) *Swarm {
 	return &Swarm{
-		ServiceManager: newServiceManager(dockerClient),
+		Services: newServiceManager(dockerClient),
+		Secrets:  newSecretManager(dockerClient),
 	}
 }
